requirements.txt
Pillow
ExifRead
ttkbootstrap
pyinstaller
